[BUGFIX] Correctly prefix resources with base path
Add tests for more complicated nested structures
to prove that it previously failed.

Main issue was root config file reader not having the
absolute path to the resource, which we now fixed
at the cost of adding a new isFileResource argument
to the factory registration, so that non file based
resources don't get their resource changed.