[!!!] Replace value with null instead of placeholder
When strict placeholder parsing is disabled, we previously
kept the placeholder in place. This however could be
quite confusing on one hand and isn't useful for consumers
on the other hand.

We now replace the placeholder with null, which more likely
matches expectations. Now config values can be checked with
isset() by consumers to verify whether a placeholder was found or not.

While this can be considered a breaking change and is marked as such,
the impact should be pretty low.