[!!!][FEATURE] Refactor placeholder processor to allow arbitrary placeholders
Instead of having an hard coded set of placeholders,
add an interface for placeholders and the possibility to
to add new/additional ones to the processor.

With this refactoring we also get consistency for recursive
palceholder resolving, which previously only worked for
the config placeholder.

At the same time we remove the public API to find placeholder
in a given array, since this isn't really useful or should rather
go into a separate class.